rename SubmissionFilesPresenter to Presenters::SubmissionFiles::Base
