+++ app/controllers/submission_files_controller.rb
@@ -4,7 +4,12 @@ class SubmissionFilesController < ApplicationController
 
     # let's use the object_stream here because there's no reason to hit S3 twice
     if presenter.submission_file_streamable?
-      send_data(presenter.stream_submission_file, filename: presenter.filename)
+      tmp_dir = Dir.mktmpdir
+      filepath = [tmp_dir, presenter.filename].join "/"
+      streamed_file = File.new(filepath, "w") {|f| f << presenter.stream_submission_file }
+      send_file streamed_file
+
+      # send_data(presenter.stream_submission_file, filename: presenter.filename)
     else
       presenter.mark_submission_file_missing
       flash[:alert] = "The requested file was not found on the server."
