--- app/views/submission_files/download.html.haml
@@ -1,1 +1,0 @@
--# leave this here for streaming data without having to circumvent render or redirect
